% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_createtable.R
\name{sql_createtable}
\alias{sql_createtable}
\title{Creates a new table within a database}
\usage{
sql_createtable(tbl_name, col_names, col_types, primary_key = -1)
}
\arguments{
\item{tbl_name}{- the name of the table to be made}

\item{col_names}{- a vector of the desired column names}

\item{col_types}{- a vector of the desired column types (length of col_names and col_types must be equal)}

\item{primary_key}{- a vector with the columns to be included in the primary key (default is all of them)}
}
\description{
Creates a new table within a database
}
\examples{
sql_createtable("Table1", c("c1","c2"), rep("INTEGER",2), primary_key = "c1")
sql_createtable("Table2", c("c1","c2", "c3"), rep("INTEGER",3), primary_key = c("c1","c2"))
sql_createtable("Table3", c("c1","c2"), rep("INTEGER",2))
}

