% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_delete.R
\name{sql_delete}
\alias{sql_delete}
\title{Deletes data from tables with SQL queries}
\usage{
sql_delete(table, where_condition = "*")
}
\arguments{
\item{table}{the table to delete from}

\item{where_condition}{a string of conditions used to specify, need to use escape key for quotes}
}
\description{
Deletes data from tables with SQL queries
}
\examples{
sql_delete("people", c("name = \\"Maria\\" OR age = 28 AND gender = \\"female\\"")
sql_delete("car", c("make = \\"Toyota\\" AND model = \\"Camry\\""))
}

