% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_droptable.R
\name{sql_droptable}
\alias{sql_droptable}
\title{Drops tables from a SQlite database}
\usage{
sql_droptable(table_name)
}
\arguments{
\item{table}{the table to delete from}
}
\description{
Drops tables from a SQlite database
}
\examples{
sql_droptable("Table_1")
}

