% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_init.R
\name{sql_init}
\alias{sql_init}
\title{Creates a global connection to a database}
\usage{
sql_init(file = -1)
}
\arguments{
\item{file}{- the name of the file to connect to}
}
\description{
Creates a global connection to a database
}
\examples{
sql_init()
sql_init("my_first_database.sqlite")

}

