% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_insert.R
\name{sql_insert}
\alias{sql_insert}
\title{Inserts data into tables with SQL queries}
\usage{
sql_insert(table, values, columns = -1)
}
\arguments{
\item{table}{- the name of the table to insert into}

\item{values}{- a vector of the values to insert}

\item{columns}{- a vector of the columns to insert into}
}
\description{
Inserts data into tables with SQL queries
}
\examples{
sql_insert("people", c("name", "sex", "age"), c("Maria", "female", "28"))
}

