% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_listtables.R
\name{sql_listtables}
\alias{sql_listtables}
\title{Function to list tables for a connection}
\usage{
sql_listtables()
}
\description{
Function to list tables for a connection
}
\examples{
sql_listtables()

}

