% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_mirror.R
\name{sql_mirror}
\alias{sql_mirror}
\title{Takes a data frame and makes it into a table in SQL}
\usage{
sql_mirror(df)
}
\arguments{
\item{df}{- the dataframe to be placed in sql}
}
\description{
Takes a data frame and makes it into a table in SQL
}

