% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_select.R
\name{sql_select}
\alias{sql_select}
\title{Runs a SQL Select query}
\usage{
sql_select(tbl, cols = "*", where = -1, orderby = -1, orderType = "ASC",
  groupBy = -1, having = -1)
}
\arguments{
\item{tbl}{- the table to run the query on}

\item{cols}{- a vector of the column names you'd like to select, default is all of them}

\item{where}{- a single string with the where conditions for the query}

\item{orderby}{- a vector of the column names you'd like to order by in the query}

\item{orderType}{- the type of ordering (ASC | DESC), DESC is default}

\item{groupBy}{-a vector of the column names you'd like to group by in the query}

\item{having}{-a single string with the having conditions for the query}
}
\value{
a dataframe with the results of the query
}
\description{
Runs a SQL Select query
}
\examples{
sql_select(table)
sqL_select(table, c("col1", "col2"), "col1>10 AND col2<10")

}

