% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_update.R
\name{sql_update}
\alias{sql_update}
\title{Updates an existing data table}
\usage{
sql_update(tbl, set, where)
}
\arguments{
\item{tbl}{- the table you'd like to perform an update on}

\item{set}{- a list (or vector/matrix/data.frame) containing vectors of values you'd like to update the selected rows with. Each different column change should be an element. Each list element is a different query.}

\item{where}{- a vector of where conditions for each of your updates}
}
\description{
Updates an existing data table
}
\examples{
sql_update("table_1", list(c("c1 = 301", "c2 =307"), c("c1 =909")), c("c1=300","c1 =99"))
}

